abgabe 07.7-2

+++ app/Http/Controllers/UserController.php
@@ -80,7 +80,10 @@ class UserController extends Controller
     {
         $user = User::find($userId);
 
-        $user->delete();
+        if ($user->id != 1) {
+            $user->delete();
+        }
+
 
         return redirect('/home');
     }
